Corrigindo tipo do campo

+++ assembleia/database/migrations/2015_08_23_010431_cria_tabelas_deputado_e_tipo_despesa_e_verba_indenizatoria.php
@@ -27,7 +27,7 @@ class CriaTabelasDeputadoETipoDespesaEVerbaIndenizatoria extends Migration
             $table->integer('codTipoDespesa')->unsigned()->default(0);
             $table->foreign('codTipoDespesa')->references('codTipoDespesa')->on('tipos_despesas')->onDelete('cascade');
             $table->text('dataReferencia')->default('');
-            $table->numeric('valor')->default('');
+            $table->float('valor')->default('');
             $table->timestamps();
         });
     }
